Cleanup and reorganization.
